CSI to use framework's fallback home activity
CSI currently uses Shell's null home activity, which is not
recognized by Settings and listed for user to select in:
  Settings > Apps & notifications > Default apps > Home app
while it is not supposed to be used by normal users.

This change lets CSI use the framework's fallback home, which
is enabled only when ro.system_user_home_needed is true and
user setup is completed.

Bug: 149307007
Bug: 149780604
Test: $ lunch csi_x86-userdebug; m
      $ lunch csi_arm64-userdebug; m
Change-Id: I87ee8e4b8238b348145ad507e0ef347198f56eab