Revert "Add virt APEX to arm64 & x86_64 GSI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/e5cc51f9-e453-42b3-9681-0292c41b2bd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe149f087f0efeece0f9f0396cd5740f18a6fe8c