Revert "Add virt APEX to arm64 & x86_64 GSI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/e6c88cc9-ad35-411e-aadf-6a64adaf352d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00c6f929bd2e87c9e52225538ce8afe5c060800a