Revert "Add virt APEX to arm64 & x86_64 GSI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/0d8dc73f-bc45-49f2-9ae7-930df214b816). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0879e93a34c4468f09c988f3cbdd991891612db4