Revert "Add virt APEX to arm64 & x86_64 GSI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6c852b9f-6807-427e-9098-0f91ad1fcfde). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia973921bdb081eb6e93188718fbd728ca09f0a8b