Revert "Add virt APEX to arm64 & x86_64 GSI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/09186e1c-e5d0-4c6d-93b6-776325729070). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I08ac67b312d9fac89a8d8a96d7f39532652083aa