Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/184f3e2b-0e0c-49f7-90cc-b7c4c2a4c49b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ff5d81fe7b3518113066f870c762f0e3d47cb83
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com